+++ changes.md
@@ -3,6 +3,7 @@
 ## ???
 * Eltako FMZ14 is working and tested ([Multifunction Time Relay](https://www.eltako.com/fileadmin/downloads/en/_bedienung/FMZ14_30014009-2_gb.pdf))
 * Adjusted device discovery to multi-gateway support
+* Device discovery detects base id of FAM14 automatically
 
 ## Version 1.2.2 Support for Multiple Gateways
 * Full support for gateway [Eltako FAM-USB](https://www.eltako.com/en/product/professional-standard-en/three-phase-energy-meters-and-one-phase-energy-meters/fam-usb/)